requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.42.1
frozenlist==1.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
httplib2==0.22.0
idna==3.4
kiwisolver==1.4.5
magic-filter==1.0.11
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.9
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
